refactor(ui): extract source input collection
